feat(i18n): Add internationalization support by restructuring pages
