--- app/(auth)/forgetPassword/page.tsx
@@ -1,5 +1,0 @@
-const ForgetPassword = () => {
-  return <div>ForgetPassword</div>;
-};
-
-export default ForgetPassword;
