+++ app/[locale]/admin/page.tsx
