+++ app/[locale]/blog/[slug]/page.tsx
