+++ app/[locale]/blog/loading.tsx
