+++ app/[locale]/create-post/page.tsx
