+++ app/[locale]/dashboard/page.tsx
